requirements.txt
cryptography
pymongo
nextcord 
python-dotenv
requests